requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
duckduckgo_search==7.5.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.0
googleapis-common-protos==1.69.2
gradio==5.21.0
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.4
openinference-instrumentation==0.1.24
openinference-instrumentation-smolagents==0.1.7
openinference-semantic-conventions==0.1.15
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.11.0
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smolagents==1.11.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
